server/requirements.txt
# Core Flask framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
cryptography==43.0.3

# API & Serialization
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
Flask-Marshmallow==0.15.0

# AI/ML Dependencies
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.25
requests==2.31.0
numpy==1.24.3
# pandas==2.1.4  # Temporarily disabled due to Python 3.13 compatibility

# HTTP Client & Async
httpx==0.26.0
aiohttp==3.9.1

# Testing Framework
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.2
factory-boy==3.3.0
faker==21.0.0

# Development Tools
black==23.12.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0

# Production/Deployment
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[flask]==1.40.0

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3

# Development
watchdog==3.0.0
flask-shell-ipython==1.4.1 